requirements.txt
llama_index
python-dotenv
openai
pyyaml
accelerate
asttokens
bitsandbytes
certifi
charset-normalizer
comm
debugpy
decorator
exceptiongroup
executing
filelock
fsspec
huggingface-hub
idna
Jinja2
MarkupSafe
matplotlib-inline
mpmath
nest-asyncio
networkx
numpy
packaging
pandas
parso
pexpect
pillow
platformdirs
psutil
ptyprocess
pure_eval
Pygments
python-dateutil
PyYAML
regex
requests
ruff
safetensors
seaborn
six
stack-data
sympy
tokenizers
torch
torchaudio
torchvision
tqdm
traitlets
transformers
triton
typing_extensions
urllib3
wcwidth
widgetsnbextension
